Fix account tests (#61)
* :truck: rename integration test

* :art: create random server instances

* :bug: solve expectation error

* :construction_worker: run integration test on pipeline
