+++ .github/workflows/account.yml
@@ -34,4 +34,4 @@ jobs:
         run: docker-compose up -d
       - name: Build with Maven
         working-directory: account
-        run: mvn -B package --file pom.xml
+        run: mvn -B verify --file pom.xml
